requirements.txt
environs==9.5.0
attrs==21.2.0
loguru==0.6.0
flask==2.0.2
aiohttp==3.8.1
requests==2.27.1
pyquery==1.4.3
redis==4.2.2
tornado==6.1
gevent==21.12.0